Fix version endian (again), bootloader size, migrate to global incomming packet_t*
Because the packet handlers for modules are called by a function pointer, it appears
that XC8 does not correctly locate the parameters passed in, as such they are
corrupted as soon as the function starts.

This does argue against this approach and perhaps it should be migrated back to
a statically linked set of functions. Using #define to include the routers for
prefixes.